--- packages/compass-collections-ddl/src/components/collection-fields/collation-fields.less
@@ -1,22 +1,0 @@
-.collation {
-  display: block;
-
-  &-field {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 15px;
-  }
-
-  &-label {
-    width: 150px;
-    text-align: right;
-    display: inline-block;
-    margin: 0 10px 0 0;
-  }
-
-  &-select {
-    width: 250px;
-    display: inline-block;
-  }
-}
