+++ packages/compass-collections-ddl/src/components/create-collection-modal/collection-fields/has-time-series-support.js
