+++ packages/compass-collections-ddl/src/components/create-collection-modal/create-collection-modal.jsx
@@ -9,7 +9,7 @@ import { createCollection } from '../../modules/create-collection';
 import { clearError } from '../../modules/error';
 import { toggleIsVisible } from '../../modules/is-visible';
 
-import CollectionFields from '../collection-fields';
+import CollectionFields from './collection-fields';
 import styles from './create-collection-modal.less';
 
 class CreateCollectionModal extends PureComponent {
@@ -63,7 +63,6 @@ class CreateCollectionModal extends PureComponent {
         buttonText="Create Collection"
         submitDisabled={this.state.submitDisabled}
         className={styles['create-collection-modal']}
-        contentClassName={styles['create-collection-modal-content']}
       >
         <CollectionFields
           serverVersion={this.props.serverVersion}
